Lifting `routeParams` to the global source definition
